test_pipeline/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyMuPDF>=1.24.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
jinja2>=3.1.0
